api_sc/requirements.txt
numpy==1.23.5
antlr4-python3-runtime==4.8
audioread==3.0.1
bitarray==2.8.3
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
Cython==3.0.5
decorator==5.1.1
fairseq==0.12.2
faiss-cpu==1.7.4
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==3.0.0
fsspec==2023.4.0
future==0.18.3
hydra-core==1.0.7
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.0
numba==0.58.1
numpy==1.23.5
omegaconf==2.0.6
packaging==23.2
Pillow==9.3.0
platformdirs==4.0.0
pooch==1.8.0
portalocker==2.8.2
praat-parselmouth==0.4.3
pycparser==2.21
python-dotenv==1.0.0
pyworld==0.3.4
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
resampy==0.4.2
sacrebleu==2.3.2
scikit-learn==1.3.2
scipy==1.11.3
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.1.1+cpu
torchaudio==2.1.1+cpu
torchcrepe==0.0.22
torchvision==0.16.1+cpu
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
zipp==3.17.0
flask-paginate==2023.10.24
Flask-Cors==4.0.0
